Bug 1136177 - Amount by mouse wheel scrolling is wrong. r=jimm

+++ widget/windows/WinMouseScrollHandler.h
@@ -289,8 +289,8 @@ private:
     bool IsPageScroll(bool aForVertical) const
     {
       MOZ_ASSERT(mInitialized, "SystemSettings must be initialized");
-      return aForVertical ? (mScrollLines == WHEEL_PAGESCROLL) :
-                            (mScrollChars == WHEEL_PAGESCROLL);
+      return aForVertical ? (uint32_t(mScrollLines) == WHEEL_PAGESCROLL) :
+                            (uint32_t(mScrollChars) == WHEEL_PAGESCROLL);
     }
 
   private:
